+++ src/main/java/com/water/quality/pojo/entity/WqParam.java
@@ -2,19 +2,20 @@ package com.water.quality.pojo.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
-import java.io.Serializable;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
  * 水质参数表
  * </p>
  *
- * @author 
+ * @author
  * @since 2022-02-10
  */
 @Data
@@ -25,13 +26,13 @@ public class WqParam implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "水质类别id")
-    private Integer wqTypeId;
+    private Long wqTypeId;
 
     @ApiModelProperty(value = "水质参数类别id")
-    private Integer wqParamTypeId;
+    private Long wqParamTypeId;
 
     @ApiModelProperty(value = "最大浓度阈值")
     private Double maxConcentration;
@@ -46,10 +47,10 @@ public class WqParam implements Serializable {
     private LocalDateTime updateTime;
 
     @ApiModelProperty(value = "创建人id")
-    private Integer creatorId;
+    private Long creatorId;
 
     @ApiModelProperty(value = "更新人id")
-    private Integer editorId;
+    private Long editorId;
 
 
 }
