+++ src/main/java/com/water/quality/pojo/entity/WqParamType.java
@@ -2,19 +2,20 @@ package com.water.quality.pojo.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
-import java.io.Serializable;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
  * 水质参数种类
  * </p>
  *
- * @author 
+ * @author
  * @since 2022-02-10
  */
 @Data
@@ -25,7 +26,7 @@ public class WqParamType implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "水质参数名称 (浑浊度，ph，溶氧量，深度值，导电率，氧化还原电位等)")
     private String name;
@@ -40,7 +41,7 @@ public class WqParamType implements Serializable {
     private LocalDateTime updateTime;
 
     @ApiModelProperty(value = "更新人id")
-    private Integer editorId;
+    private Long editorId;
 
 
 }
